List Exercise

+++ module-1/07_Collections_Part_1/student-exercise/Exercises/05_ArrayInt2ListDouble.cs
@@ -19,11 +19,10 @@ namespace Exercises
             
             List<double> dividedTwo = new List<double>();
             
-            
             {
                 for (int i = 0; i < intArray.Length; i++)
                 {
-                    dividedTwo.Add(intArray[i] / 2);
+                    dividedTwo.Add(intArray[i] / (double)2);
                 }
             }
